requirements.txt
streamlit
ccxt
pandas
python-telegram-bot